requirements.txt
requests
python-dotenv
qdrant-client
groq
